+++ src/util/parseAlgoliaHit.js
